ai-server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# STT
openai-whisper==20231117

# Audio processing
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4

# ML / Acoustic emotion
torch==2.3.0
torchaudio==2.3.0
transformers==4.41.0

# Text sentiment
sentencepiece==0.2.0

# Utility
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
